Updates: Successfully applied the generation process of start and end date

+++ src/main/java/sys/app/csc/repository/ProjectPlanRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import sys.app.csc.entity.ProjectPlanEntity;
 
 @Repository
-public interface ProjectPlanRepository extends JpaRepository<ProjectPlanEntity, Long> {
+public interface ProjectPlanRepository extends JpaRepository<ProjectPlanEntity, Integer> {
 
 }
